requirements.txt
Flask
coloredlogs
PyGithub
authlib
Flask-Login
flask_dance
